hybrid-trading-simulator/backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
aio-pika
motor
python-dotenv
web3